+++ graphql-maven-plugin-logic/src/test/java/com/graphql_java_generator/plugin/compilation_tests/AllGraphQLCasesServer_separateUtilityClasses_Test.java
@@ -19,6 +19,6 @@ class AllGraphQLCasesServer_separateUtilityClasses_Test extends AbstractIntegrat
 	@BeforeEach
 	public void setUp() {
 		((GraphQLConfigurationTestHelper) pluginConfiguration).separateUtilityClasses = true;
-		graphqlTestHelper.checkSchemaStringProvider("allGraphQLCases.graphqls");
+		graphqlTestHelper.checkSchemaStringProvider("allGraphQLCases*.graphqls");
 	}
 }