+++ graphql-maven-plugin/src/main/java/com/graphql_java_generator/mavenplugin/GenerateRelaySchemaMojo.java
@@ -39,9 +39,16 @@ public class GenerateRelaySchemaMojo extends AbstractMojo {
 	String schemaFileName;
 
 	/**
+	 * <P>
 	 * The pattern to find the graphql schema file(s). The default value is "/*.graphqls" meaning that the maven plugin
 	 * will search all graphqls files in the "/src/main/resources" folder (please check also the <I>schemaFileFolder</I>
 	 * plugin parameter).
+	 * </P>
+	 * <P>
+	 * You can put the star (*) joker in the filename, to retrieve several files at ones, for instance
+	 * <I>/myschema*.graphqls</I> will retrieve the <I>/src/main/resources/myschema.graphqls</I> and
+	 * <I>/src/main/resources/myschema_extend.graphqls</I> files.
+	 * <P>
 	 */
 	@Parameter(property = "com.graphql_java_generator.mavenplugin.schemaFilePattern", defaultValue = GenerateRelaySchemaConfiguration.DEFAULT_SCHEMA_FILE_PATTERN)
 	String schemaFilePattern;